Created the address controller & service

+++ src/address/address.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, NotFoundException} from '@nestjs/common';
 import { AddressService } from './address.service';
 import { CreateAddressDto } from './dto/create-address.dto';
 import { UpdateAddressDto } from './dto/update-address.dto';
@@ -13,22 +13,38 @@ export class AddressController {
   }
 
   @Get()
-  findAll() {
-    return this.addressService.findAll();
+  async findAll() {
+    const addresses = await this.addressService.findAll();
+    if (!addresses || addresses.length === 0) {
+      throw new NotFoundException('No addresses found');
+    }
+    return addresses;
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.addressService.findOne(+id);
+  async findOne(@Param('id') id: string) {
+    const address = await this.addressService.findOne(+id);
+    if (!address) {
+      throw new NotFoundException(`Address with ID ${id} not found`);
+    }
+    return address;
   }
 
   @Patch(':id')
   update(@Param('id') id: string, @Body() updateAddressDto: UpdateAddressDto) {
+    const address = this.addressService.findOne(+id);
+    if (!address) {
+      throw new NotFoundException(`Address with ID ${id} not found`);
+    }
     return this.addressService.update(+id, updateAddressDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  async remove(@Param('id') id: string) {
+    const address = await this.addressService.findOne(+id);
+    if (!address) {
+      throw new NotFoundException(`Address with ID ${id} not found`);
+    }
     return this.addressService.remove(+id);
   }
 }
